+++ packages/RSuite/NEWS.md
@@ -1,8 +1,10 @@
-# RSuite 0.33 (2018-08-??)
+# RSuite 0.33 (2018-09-??)
   * Detection of implicit dependencies like <pkg>::<name> or <pkg>:::<name> in
     master scripts (and tests) added.
   * Added R Suite project templates to RStudio project templates. We can now create
     R Suite projects using "File -> New Project..." in RStudio 
+  * Caching of in-src-archive packages list implemented.
+  * Got rid of subprocess in favor to processx package.
 
 # RSuite 0.32 (2018-08-15)
   * Unregistering repo adapters functionality added.